US12-Middleware: tests for the middleware have been added, passed, and all functionality verified. This thing is good to go. All related tasks should be complete.

+++ Psyche VR Experience/Assets/Scripts/DB/ArtworkData.cs
@@ -13,5 +13,5 @@ public class ArtworkData : ScriptableObject
     [SerializeField] public String artworkDate;                 //MONTH - DAY - YEAR (As seen on Website)
     [SerializeField] public List<String> artworkURLs;           //urls for artwork
     [SerializeField] public int artworkCount;                   //number of pieces of artwork
-    [SerializeField] public int artworkID;                      //ID number of the artwork (This is for our use)
+    [SerializeField] public long artworkID;                      //ID number of the artwork (This is for our use)
 }
